Gnus: Group: Only check groups under certain level but don't relist
I want to have groups which I only want to fetch when I explicitly
choose to do so but not always. Groups like these are above the level
of Groups I want to fetch news for.
But then I don't want to only list the groups at or below the
get-new-news-level after getting new news.

Add to advice functions to archive this:
- The first advice filter the get new news command to use
  the fetch level I defined unless the function gets an explicit
  level. This ensures the calling the function which an explicit
  level of groups to get still works.
- Advice number two workarounds the side of effect advice one
  and of the general behavior of the get news function to call
  the list group function which hides all the groups below the
  fetch level. The advice checks if the caller or command
  is gnus-group-get-new-news to not update the group-level.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>