Gnus: Build sparse threads and fetch all old headers
I did not know that the option was not the default. Set it to 'more
it could slow down the loading or building of summaries when entering
groups but it is very useful to have all reference to threads.

Also fetch all old header to help search for old messages when
requesting to load older messages to load a thread.

Some values might turn out to slow down to far.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>