Gnus: Art: Fix fontification of citations
The reason for the issue in citations was that
`gnus-treat-highlight-citation' was set to `'last'.
Valid values for the variable are `t' or a number even thou
first works too.

Setting it to the default `t' fixes the fontification of citations.
Note that without `gnus-unbuttonized-mime-types' set to
`'("text/plain")' only the first parts citations are fontified
correctl, the default of the variable unbuttonizes all types.

The related Emacs bug is: #66779.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>