Borg: Sync drone urls: Update for recent forge and properly connection type into account
forge--url-equal got removed. The url comparison had to be refactored.
While doing so it made sense to split the default url into plain SSH
syntax or URI SSH syntax so it can be compared correctly.
I.e. cases where the URL is the same but Epkg returns it as HTTPS
while we might use SSH.
Also at the same time take the SSH URL as new URL in case the previous
URL was also one.