Disable the creation of ID's for this file
We want to create ID's for todos which don't have ID's.
However not for this file as it isn't related to
the agenda and links to it should be stable
without adding ID's, as they user specific.

Mark the new setting of this variable as safe in init.el
since users of it should of course should assume it as safe
as it comes from the same place init.el comes from.